requirements.txt
aiokafka==0.8.1
aiocache==0.12.2
aiomqtt==1.2.1
dataclasses-avroschema
httpx==0.25.0
python-dotenv==1.0.0
sentry-sdk==1.17.0
typing_extensions==4.8.0
orjson
python-dotenv
